% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.r
\name{calculate_gtd_pr}
\alias{calculate_gtd_pr}
\title{Calculate precision, recall and F1 score for an edgelist}
\usage{
calculate_gtd_pr(
  e,
  min_weight = NA,
  weight_col = "weight",
  hourdiff_range = c(0, Inf),
  gtd_id_filter = NULL
)
}
\arguments{
\item{e}{An edgelist in which the from column contains GTD event ids, and the to column contains
The Guardian news article ids}

\item{min_weight}{Optionally, specify minimum weight to filter e before calculation}
}
\value{
A list with P, R and F1 scores for both the match and article level.
}
\description{
This function calculates the precision, recall and F1 score for an edgelists containing matches
of GTD events and news articles from The Guardian
}
\details{
See gold_matches$description for more information on the gold standard
}
