% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{country_name_matches}
\alias{country_name_matches}
\title{Matches of different country names in GTD data and UN votes}
\format{
A data.table
}
\usage{
data(country_name_matches)
}
\description{
Matches of different country names in GTD data and UN votes
}
\examples{
head(country_name_matches)
}
\keyword{datasets}
