% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.r
\name{download_guardian}
\alias{download_guardian}
\title{Download articles from the Guardian}
\usage{
download_guardian(
  query_terms,
  api.key,
  fromdate,
  todate,
  skip = NULL,
  path = getwd(),
  stepsize = 7,
  verbose = T
)
}
\arguments{
\item{query_terms}{A character vector with Guardian queries. All terms are concatenated with OR operators
(grouped in parentheses).}

\item{api.key}{An API key for the Guardian API, which can be obtained for free by filling in
\href{https://bonobo.capi.gutools.co.uk/register/developer}{this webform}.}

\item{fromdate}{The starting date, in format "YYYY-MM-DD". e.g. "2010-01-01"}

\item{todate}{The ending date, in format "YYYY-MM-DD". e.g. "2015-12-31"}

\item{skip}{Optionally, a vector of days to skip, as a last resort.}

\item{path}{The path to a directory in which the download directory is created.
Default is current working directory}

\item{stepsize}{The number of days per batch.}

\item{verbose}{If TRUE, report progress}
}
\value{
A data.frame with guardian articles.
}
\description{
This is a wrapper for the \code{\link[GuardianR]{get_guardian}} function from the GuardianR package.
Given a query, or character vector of query terms, and a from and to date, it will download the data
in batches, and once downloaded return all articles as a data.frame. This makes it possible to collect
all the data. If the download is interrupted (e.g., internet issue, leaving work, tornado), it can
simply be resumed  by running the function again with the same arguments.
}
\examples{
\donttest{
api.key = "[your own api key]"
query_terms = terrorism_news    ## (terrorism_news is data included in this package)
d = download_guardian(query_terms, api.key, fromdate = '2012-01-01', todate='2012-01-20')
}
}
