% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoplot.r
\name{get_region_id}
\alias{get_region_id}
\title{Given longitude and latitude, return region id as used by ggmap}
\usage{
get_region_id(lon, lat, country = NULL, radius = 10)
}
\arguments{
\item{lon}{longitude}

\item{lat}{lattitude}

\item{country}{If country names are available, and these match regions in the
world map data, give these matches preferences over nearest neigbour matches}

\item{radius}{max radius in which to look}
}
\value{
A character vector with the region names as used by ggmap
}
\description{
Finds nearest nabour of lon/lat for regions in the world map data in ggmap
}
\examples{

}
