% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gold_matches}
\alias{gold_matches}
\title{Gold standard for GTD event matches}
\format{
A list with three slots: $description, $articles and $matches
}
\usage{
data(gold_matches)
}
\description{
Gold standard for GTD events matches in 500 random articles from the Guardian between
2000 and 2015. Articles were manually coded, by looking whether a new terrorist event
was reported in the headline and lead (i.e. was the main topic of the news). If so, all
GTD events within 7 days before the article publication date and in the same country as
mentioned in the article were manually checked for a match. All matches are reported in
the $matches slot, which contains pairs of guardian article ids and GTD event ids."
}
\examples{
gold_matches$description
head(gold_matches$articles)
head(gold_matches$matches)
}
\keyword{datasets}
