% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.r
\name{gtd_compare_pr}
\alias{gtd_compare_pr}
\title{Compare results of gtd_pr}
\usage{
gtd_compare_pr(pr1, pr2)
}
\arguments{
\item{pr1}{Results of gtd_pr}

\item{pr2}{Results of gtd_pr}
}
\value{
A plot an a list with data.frames
}
\description{
Calculates the difference of the precision and recall scores in pr1 and pr2.
This is calculated as pr2 minus pr1. So if precision in pr2 is 20, and precision in pr1 is 15,
the difference indicates an improvement of 5.
}
