% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.r
\name{gtd_pr}
\alias{gtd_pr}
\title{Plot Precision and Recall scores}
\usage{
gtd_pr(
  g,
  weight_range,
  steps,
  weights = NULL,
  weight_col = "weight",
  filter = NULL,
  gtd_id_filter = NULL
)
}
\arguments{
\item{g}{An edgelist, as created with newsflow.compare}

\item{weight_range}{The range of weight to plot on the x axis}

\item{steps}{The number of ticks on the x axis}

\item{weights}{Optionally, provide a vector of weights. This overrides the weight_range and steps arguments}

\item{weight_col}{the name of the column with the weight scores}

\item{filter}{Optionally, a logical vector indicating which rows of g$d to use.}
}
\value{
A plot an a list with data.frames
}
\description{
Uses the gold standard in the gold_matches data to calculate Precision and Recall scores.
Results are calculated separately for the article level (does the article have at least one GTD match) and
match level (each individual match of an article to a GTD event). Results are also calculated for different
Similarity thresholds, which will be plotted on the x-axis.
}
\details{
The results are also returned as a list with data.frames.
See gold_matches$description for details on the gold standard.
}
