% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoplot.r
\name{plot_worldmap}
\alias{plot_worldmap}
\title{Plot values on worldmap for given regions}
\usage{
plot_worldmap(region, value, scale = F, sqrt = F, only_europe = F)
}
\arguments{
\item{region}{The region id (see get_region_id function)}

\item{value}{A numeric value}

\item{scale}{if TRUE, scale data so that middle is zero}

\item{sqrt}{If TRUE, perform square root transformation. This is usefull for
pulling back outliers that would otherwise make the color differences
difficult to see}
}
\value{

}
\description{
Plot values on worldmap for given regions
}
