% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_validation.r
\name{precision_validation_sample}
\alias{precision_validation_sample}
\title{Extract a sample of edges to validate precision}
\usage{
precision_validation_sample(
  g,
  min_weight,
  n = 100,
  min_daydiff = 0,
  max_daydiff = 6
)
}
\arguments{
\item{g}{The RNewsflow edgelist}

\item{min_weight}{The weight threshold for when an edge should be considered a match}

\item{n}{The number of matches to sample}

\item{min_daydiff}{The minimum number of days between the event and the news article}

\item{max_daydiff}{The maximum number of days between the event and the news article}
}
\value{

}
\description{
Makes it easier to inspect whether matches past a certain similarity threshold (positives) are correct matches (true positives)
}
