% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.r
\name{prepare_gtd}
\alias{prepare_gtd}
\title{Prepare the GTD data}
\usage{
prepare_gtd(d, fromdate = NULL, todate = NULL, rm_time = T)
}
\arguments{
\item{d}{A data.frame with the GTD data. You can download the GTD data for free
by filling out \href{https://www.start.umd.edu/gtd/contact/}{this form}.
You can read the .xlsx file with \code{\link[openxlsx]{read.xlsx}}.}

\item{fromdate}{A Date (or character value that can be coerced to Date)}

\item{todate}{A Date (or character value that can be coerced to Date)}

\item{rm_time}{If TRUE, remove time related terms, such as monday, januari and yesterday.}
}
\value{
A quanteda style DTM
}
\description{
Creates a document term matrix in the \code{\link[quanteda]{dfm}} class.
Includes only the meta data used for the analysis presented in the vignette.
}
\examples{
\donttest{
gtd_data = openxlsx::read.xlsx('GTD FILE AS DOWNLOADED FROM GTD WEBSITE')
gtd_dtm = prepare_gtd(gtd_data, fromdate='2010-01-01')
}
}
