% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.r
\name{prepare_news}
\alias{prepare_news}
\title{Prepare news data}
\usage{
prepare_news(
  d,
  doc_col,
  date_col,
  text_cols,
  docvars = NULL,
  first_n_words = NA,
  rm_time = T
)
}
\arguments{
\item{d}{A data.frame in which each row is a news article.}

\item{doc_col}{The name of the column with unique article ids}

\item{date_col}{The name of the column with the date, in a format that can be coerced to a Date object with as.Date()}

\item{text_cols}{The name(s) of the column(s) with article text. e.g. c('headline','byline','body')}

\item{docvars}{The columns to include as docvars (article meta such as headline, url)}

\item{first_n_words}{Optionally, only include the first n words of the article text in the DTM}

\item{rm_time}{If TRUE, remove time related terms, such as monday, januari and yesterday.}
}
\value{
A quanteda style DTM
}
\description{
Prepare news data
}
