% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{terrorism_terms}
\alias{terrorism_terms}
\title{Terrorism terms}
\format{
A character vector
}
\usage{
data(terrorism_terms)
}
\description{
A set of query terms for finding news items that potentially mention terrorist events.
The list is intentially very broad. The goal is to have high recall (i.e. find all events
that mention terrorist events) and we are not that concerned about precision (i.e. finding only
events that mention terrorist events) in the data collection phase.
}
\examples{
terrorism_terms
}
\keyword{datasets}
