% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{UK_dist_scores}
\alias{UK_dist_scores}
\title{The political proximity, value distance and geographic distance of the UK to other countries}
\format{
A data.table

A data.table
}
\usage{
data(UK_dist_scores)

data(UK_geo_dist)
}
\description{
The political proximity, value distance and geographic distance of the UK to other countries

The geographic distance of UK to other countries. Contains more countries than UK_dist_scores because it does not contain the cultural distances
}
\examples{
head(UK_dist_scores)
head(UK_geo_dist)
}
\keyword{datasets}
